scripts/requirements.txt
python-dotenv
supabase
voyageai
